Fixed bugs for creating pool

+++ src/valueswap_backend/src/vault/lp_tokens.rs
@@ -7,15 +7,15 @@ use crate::utils::types::*;
 // use crate::vault::vault_pool::VAULT;
 
 thread_local! {
-    static TOTAL_LP_SUPPLY : RefCell<f64> = RefCell::new(0.0);
-    static POOL_LP_SHARE : RefCell<HashMap<String , f64>> = RefCell::new(HashMap::new());
+    static TOTAL_LP_SUPPLY : RefCell<u64> = RefCell::new(0);
+    static POOL_LP_SHARE : RefCell<HashMap<String , u64>> = RefCell::new(HashMap::new());
 }
 
 // To map pool with their LP tokens
 #[query]
-fn pool_lp_tokens(params: CreatePoolParams) -> HashMap<String, f64> {
+fn pool_lp_tokens(params: CreatePoolParams) -> HashMap<String, u64> {
     POOL_LP_SHARE.with(|lp_share| {
-        let pool_supply: f64 = params
+        let pool_supply: u64 = params
             .values
             .iter()
             .zip(params.balances.iter())
@@ -31,10 +31,10 @@ fn pool_lp_tokens(params: CreatePoolParams) -> HashMap<String, f64> {
 
 // To get all lp tokens
 #[query]
-fn total_lp_tokens() -> f64 {
-    let mut total_supply: f64 = 0.0;
+fn total_lp_tokens() -> u64 {
+    let mut total_supply: u64 = 0;
     POOL_LP_SHARE.with(|share| {
-        let temp: HashMap<String, f64> = share.borrow().clone();
+        let temp: HashMap<String, u64> = share.borrow().clone();
         for (_key, value) in temp.iter() {
             total_supply += value
         }
@@ -47,9 +47,9 @@ fn total_lp_tokens() -> f64 {
 
 // Query to get LP tokens for a specific pool
 #[query]
-fn get_lp_tokens(pool_name: String) -> Option<f64> {
+fn get_lp_tokens(pool_name: String) -> Option<u64> {
     POOL_LP_SHARE.with(|share| {
-        let temp: HashMap<String, f64> = share.borrow().clone();
+        let temp: HashMap<String, u64> = share.borrow().clone();
         if let Some(key) = temp.get(&pool_name) {
             Some(*key)
         } else {
